Volume decrease is more natural, but still has popping
